refactor(docs): walkMdx 共通化・DummyText を UI Components に統一・未使用 import 削除
- util.ts に walkMdx(dir, { skipUnderscore? }) を追加し、build-llms-txt の
  listMdxFiles と build-ui-index-md の listUiMdx を統合
- build-llms-txt: DummyText を Optional から外し UI Components に分類
- build-ui-index-md: 未使用の HastRoot import を削除
- build-llms-txt.test.ts: DummyText の分類テストを UI Components に更新
